Add `object.copy` builtin
The main motivation for this is to make it easier to perform
matrix-style testing. Without this, mutating a fixture stored in data
has unpredictable results. Here is an example of code affected by this
(both tests pass).

```rego
thing = input.thing.stuff

test_thing {
    0 == thing with input.thing as data.fixture
    5 == thing with input.thing as data.fixture with input.thing.stuff as 5
}

test_thing_2 {
    5 == thing with input.thing as data.fixture
}
```

The "as" call results in a deep mutation of the original fixture, with
unpredictable results for every test that follows that uses the same
fixture.

Signed-off-by: Mike Lundy <mike@fluffypenguin.org>